working on matcap cube. put suspense in canvas
