+++ package.json
@@ -9,6 +9,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@react-three/drei": "^9.57.3",
     "@react-three/fiber": "^8.12.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
