+++ src/components/About.tsx
@@ -1,10 +1,11 @@
 import React from "react";
+import MyCanvas from "./canvas/MyCanvas";
 
 function About() {
   return (
     <div>
       <h2>Who is 「Rose」?</h2>
-      <p></p>
+      <MyCanvas />
     </div>
   );
 }
