+++ src/components/Homepage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import MyCanvas from "./MyCanvas";
+import MyCanvas from "./canvas/MyCanvas";
 
 function Homepage() {
   return (
