+++ tailwind.config.cjs
@@ -3,7 +3,7 @@ module.exports = {
   // If you want to support toggling dark mode manually instead of relying on the operating system preference,
   // use the class strategy instead of the media strategy.
   darkMode: "class",
-  content: ["./index.html", "./src/**/*.{js,ts,jsx,tsx}"],
+  content: ["./index.html", "./src/**/*.{html,js,ts,jsx,tsx}"],
   theme: {
     container: {
       center: true,
